AI_Knowledge_assistant/AI_Knowledge_Assistant/backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
openai
